Created new project list:
- segregated project in different components.
