+++ src/components/person-list/person-age/PersonAge.js
@@ -1,0 +1,10 @@
+import React from "react";
+import '../person-age/PersonAge.css';
+
+const PersonAge = (props) => {
+    return (
+        <div>{props.age}</div>
+    )
+}
+
+export default PersonAge;
