+++ src/components/person-list/person-date/PersonDate.css
@@ -1,0 +1,5 @@
+.date {
+    border: #222222 1px solid;
+    border-radius: 10px;
+    width: 10%;
+}
