+++ src/components/person-list/person-date/PersonDate.js
@@ -1,0 +1,19 @@
+import React from "react";
+import '../person-date/PersonDate.css';
+
+const PersonDate = (props) => {
+    const month = props.date.toLocaleString('en-US', {month: 'long'});
+    const day = props.date.toLocaleString('en-US', {day: '2-digit'});
+    const year = props.date.getFullYear();
+
+    return (
+        <div className='date'>
+            <div>{month}</div>
+            <div>{day}</div>
+            <div>{year}</div>
+        </div>
+    );
+
+
+}
+export default PersonDate;
