+++ src/components/person-list/person-name/PersonName.css
@@ -1,0 +1,5 @@
+.person__name {
+    color: red;
+    font-size: large;
+
+}
