+++ src/components/person-list/person.css
@@ -2,9 +2,9 @@
     border: blueviolet 2px solid;
     background-color: #bfd999;
     margin: 10px;
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+
 }
 
-.person__wrapper {
-    display: flex;
-    justify-content: space-around;
-}
